requirements.txt
asgiref==3.4.0
async-generator==1.10
attrs==22.1.0
autopep8==1.5.7
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.6.15
cffi==1.15.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.4
django-bulma==0.8.2
django-crispy-forms==1.13.0
django-import-export==2.8.0
django-mathfilters==1.0.0
django-pdfkit==0.3.1
django-rest-framework==0.1.0
django-widget-tweaks==1.4.12
djangoajax==3.3
djangorestframework==3.13.1
et-xmlfile==1.1.0
greenlet==1.1.3
h11==0.13.0
idna==3.3
MarkupPy==1.14
numpy==1.23.2
odfpy==1.4.1
openpyxl==3.0.10
outcome==1.2.0
pandas==1.4.3
pdfkit==0.6.0
protobuf==3.20.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
pycodestyle==2.7.0
pycparser==2.21
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2021.1
PyYAML==6.0
schedule==1.1.0
selenium==4.4.3
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.40
sqlparse==0.4.1
tablib==3.2.1
toml==0.10.2
trio==0.21.0
trio-websocket==0.9.2
tzdata==2022.2
urllib3==1.26.12
websocket-client==1.3.3
whitenoise==6.2.0
wsproto==1.2.0
xlrd==2.0.1
xlwt==1.3.0
yliveticker==0.3.0
